[refactor] rename pkg signature -> sig
