+++ txn/sig/sighash/SigHashFlag.go
